% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_package_admin.R
\docType{package}
\name{simid.foreach}
\alias{simid.foreach}
\alias{simid.foreach-package}
\title{simid.foreach: Extended foreach tools.}
\description{
simid.foreach: Extended foreach tools.
}
\author{
Lander Willem [lander.willem@uantwerpen.be]
}
\keyword{internal}
