% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_copyright.R
\name{smd_add_copyright}
\alias{smd_add_copyright}
\title{Include (C) SIMID in plot}
\usage{
smd_add_copyright(text.cex = 0.4)
}
\arguments{
\item{text.cex}{the size of the text (default = 0.4)}
}
\description{
Add a SIMID copyright statement to the bottom right corner of your figure
}
\examples{
# example 1 (default)
plot(1)
smd_add_copyright()

# example 2 (different scales)
plot(c(-1e6,1e5),c(1e4,1e5))
smd_add_copyright()

plot(1:10)
smd_add_copyright(text.cex = 1)

}
