% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_description_file.R
\name{smd_create_meta_data_file}
\alias{smd_create_meta_data_file}
\title{Create a text file with meta info on a model run}
\usage{
smd_create_meta_data_file(output_dir, run_tag = NA, run_time = NA,
  root_folder = ".")
}
\arguments{
\item{output_dir}{the folder with model output, where the new file should be created}

\item{run_tag}{a tag to specify the model run (optional)}

\item{run_time}{the run time (optional)}

\item{root_folder}{the root folder of the model software (default = '.')}
}
\description{
Create a file with meta information on current software state and run info in the given output directory.
}
