% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_metadata_file.R
\name{smd_create_metadata_file}
\alias{smd_create_metadata_file}
\title{Create a text file with metadata}
\usage{
smd_create_metadata_file(output_dir, file_prefix = "", run_tag = NA,
  run_time = NA, root_dir = ".", other_info = NA)
}
\arguments{
\item{output_dir}{the directory where the new file should be created}

\item{run_tag}{the tag to specify the model run (optional)}

\item{run_time}{the run time (optional)}

\item{root_dir}{the root directory, with a DESCRIPTION OR METADATA file to start from (default = '.')}

\item{other_info}{additional info to include in the metafile (optional)}
}
\description{
Create a file with metadata on e.g. current software state, run info, and other info.
}
