% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_file_admin.R
\name{smd_file_path}
\alias{smd_file_path}
\title{Get (and create) folder path}
\usage{
smd_file_path(..., .verbose = TRUE)
}
\arguments{
\item{...}{the directory name(s)}

\item{.verbose}{boolean to print a message if a new folder is created (default: TRUE)}
}
\description{
Get (and create) folder path
}
\keyword{external}
