% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simid_foreach.R
\name{smd_foreach}
\alias{smd_foreach}
\title{Extended foreach function}
\usage{
smd_foreach(...)
}
\arguments{
\item{...}{many parameters.}
}
\description{
Parallel processing with status bar
}
\keyword{external}
