% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cli.R
\name{smd_get_console_color}
\alias{smd_get_console_color}
\title{Get default Rstudio console color}
\usage{
smd_get_console_color(WARNING = FALSE)
}
\description{
The default color of the console (or terminal) depends on whether R(studio)
is used in DARK mode. This function return 'black' or 'white', based on the current theme.
If the function is used in R, an empty color is returned.
}
\keyword{internal}
