% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_package_version.R
\name{smd_get_local_git_commit_tag}
\alias{smd_get_local_git_commit_tag}
\title{Get the local git commit id and date}
\usage{
smd_get_local_git_commit_tag()
}
\description{
Get the local git commit id and date
}
