% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_package_version.R
\name{smd_increment_package_version_number}
\alias{smd_increment_package_version_number}
\title{Update the version number in DESCRIPTION}
\usage{
smd_increment_package_version_number(packageLocation = ".")
}
\arguments{
\item{packageLocation}{the location of the package.}
}
\description{
Increment the version number if the git commit number in the DESCRIPTION file does not correspond with the latest git commit number in the repository.
}
