% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_xml.R
\name{smd_listToXML}
\alias{smd_listToXML}
\title{Convert a list into XML format (RECURSIVE FUNCTION)}
\usage{
smd_listToXML(node, sublist)
}
\arguments{
\item{node}{the XML node}

\item{sublist}{the list to be stored}
}
\description{
The counterpart of the "XML::xmlToList" function
}
