% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cli.R
\name{smd_progress}
\alias{smd_progress}
\title{Print progress bar [DEPRECATED]}
\usage{
smd_progress(
  i_current,
  i_total,
  time_stamp_loop = Sys.time(),
  par_nodes_info = NA
)
}
\description{
please use smd_print_progress()
}
