% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_file_admin.R
\name{smd_read_rdata}
\alias{smd_read_rdata}
\title{Reads a file in RData format and returns the stored variable.}
\usage{
smd_read_rdata(filename)
}
\arguments{
\item{filename}{The filename of the .RData File (case sensitive!)}
}
\description{
Reads a file in RData format and returns the stored variable.
}
\note{
This function enables to rename a stored variable in the (compressed)
RData format. If the file contains multiple variables, this functions returns NULL.
}
\keyword{external}
