% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_parallel_admin.R
\name{smd_start_cluster}
\alias{smd_start_cluster}
\title{Start parallel working nodes}
\usage{
smd_start_cluster(.env = .GlobalEnv)
}
\description{
Start parallel working nodes
}
\note{
Pass .GlobalEnv to each node
}
\keyword{external}
