% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_package_admin.R
\docType{package}
\name{simid.rtools}
\alias{simid.rtools}
\title{simid.rtools: Extended R functions.}
\description{
Package with common R functions for different projects within SIMID.
}
\author{
Lander Willem [lander.willem@uantwerpen.be]
}
\keyword{internal}
