% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_packages.R
\name{smd_load_packages}
\alias{smd_load_packages}
\title{Load the given packages (and install them if not present yet)}
\usage{
smd_load_packages(all_packages)
}
\description{
Load the given packages (and install them if not present yet)
}
\keyword{external}
