% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cli.R
\name{smd_print}
\alias{smd_print}
\title{Print (warning) message to Console}
\usage{
smd_print(..., WARNING = F, FORCED = F)
}
\arguments{
\item{...}{(parts of the) message to print}

\item{WARNING}{boolean, to print the message in red}

\item{FORCED}{boolean, to print the message irrespectively of the parallel thread}
}
\description{
Command line interface: print message with time tag, if called
by the master node or first parallel slave
}
\keyword{internal}
