% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cli.R
\name{smd_progress}
\alias{smd_progress}
\title{Print progress bar}
\usage{
smd_progress(i_current, i_total, time_stamp_loop = Sys.time())
}
\arguments{
\item{i_current}{current iteration}

\item{i_total}{total iterations}

\item{time_stamp_loop}{starting time}
}
\description{
Progress bar interface
}
\keyword{internal}
