% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_xml.R
\name{smd_save_as_xml}
\alias{smd_save_as_xml}
\title{Save a list in XML format with given root node}
\usage{
smd_save_as_xml(data_list, root_name, file_name_prefix = "data",
  xml_prefix = NULL)
}
\arguments{
\item{data_list}{the list to be stored}

\item{root_name}{the name of the XMP main rootd}

\item{file_name_prefix}{the prefix for the file name}

\item{xml_prefix}{the xml prefix tag (optional)}
}
\description{
The counterpart of the "XML::xmlToList" function
}
\keyword{external}
