% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_parallel_admin.R
\name{smd_start_cluster}
\alias{smd_start_cluster}
\title{Start parallel working nodes}
\usage{
smd_start_cluster(timeout = 100, num_proc = NA)
}
\arguments{
\item{timeout}{The timeout in seconds for the nodes}

\item{num_proc}{The number of parallel workers to start (NA = default, number of CPU cores)}
}
\value{
Cluster info, also available as global variable 'par_nodes_info'
}
\description{
Start parallel working nodes
}
\note{
Explicitly setup the PSOCK cluster using the 'sequential' setup strategy, to prevent an error in RStudio on macOS with R 4.0.x (https://github.com/rstudio/rstudio/issues/6692)
}
\keyword{external}
