% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_parallel_admin.R
\name{smd_start_cluster}
\alias{smd_start_cluster}
\title{Start parallel working nodes}
\usage{
smd_start_cluster(timeout = 100)
}
\arguments{
\item{timeout}{The timeout in seconds for the nodes}
}
\description{
Start parallel working nodes
}
\keyword{external}
